refactor[]:rest request adapter added
