+++ HealthcareSystem/PatientWebApp/Controllers/CityController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
+using PatientWebApp.Controllers.Adapter;
 using PatientWebApp.Settings;
 using RestSharp;
 
@@ -22,16 +23,7 @@ namespace PatientWebApp.Controllers
         [HttpGet("country/{countryId}")]
         public IActionResult GetCitiesByCountryId(int countryId)
         {
-            var client = new RestClient(_serviceSettings.UserServiceUrl);
-            var request = new RestRequest("/api/city/country/" + countryId);
-            var response = client.Execute(request);
-            var contentResult = new ContentResult();
-
-            contentResult.Content = response.Content;
-            contentResult.ContentType = "application/json";
-            contentResult.StatusCode = (int)response.StatusCode;
-
-            return contentResult;
+            return RequestAdapter.SendGetRequest(_serviceSettings.UserServiceUrl, "/api/city/country/" + countryId);
         } 
     }
 }
