--- HealthcareSystem/PatientWebApp/DTOs/UserCredentialsDTO.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace PatientWebApp.DTOs
-{
-    public class UserCredentialsDTO
-    {
-        public string Username { get; set; }
-        public string Password { get; set; }
-    }
-}
